+++ packages/terradart_google/lib/project.dart
@@ -2,7 +2,6 @@
 /// Project services (enabling individual GCP APIs on a project).
 library;
 
-export 'src/project/api_enablement.dart' show ApiEnablement, ApisEnablement;
 export 'src/project/apis.dart' show Apis;
 export 'src/project/barrels.dart' show Barrels;
 export 'src/project/google_project_service.dart' show GoogleProjectService;