+++ packages/terradart_google/test/time/time_sleep_test.dart
@@ -1,7 +1,8 @@
 import 'package:terradart_core/terradart_core.dart';
+import 'package:terradart_google/time.dart';
 import 'package:test/test.dart';
 
-import 'helpers/fake_resources.dart';
+import '../_helpers.dart';
 
 void main() {
   group('TimeSleep', () {